server/requirements.txt

Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
pathlib==1.0.1
six==1.16.0
django-nose==1.4.7